+++ test/bigtest/tests/routes/items-route-test.js
@@ -1,16 +1,15 @@
 import { beforeEach, describe, it } from '@bigtest/mocha';
-
 import { expect } from 'chai';
 
-import setupApplication from '../../../helpers/setup-application';
-import ItemsRouteInteractor from '../../../interactors/routes/items-route';
+import setupApplication from '../../helpers/setup-application';
+import ItemsRouteInteractor from '../../interactors/routes/items-route';
 
 describe('ItemsRoute', () => {
   setupApplication();
 
   const itemsRoute = new ItemsRouteInteractor();
 
-  beforeEach(async function () {
+  beforeEach(function () {
     this.visit('/inventory/items');
   });